구조 수정
. 이차원 벡터 m_height의 타입을 BYTE 대신 FLOAT로 수정
. Object 및 Camera에서 셰이더로 리소스를 전송할 때 루트 상수 대신 루트 서술자를 이용하도록 수정
- 그에 따른 UploadBuffer 템플릿 클래스 생성.

+++ Computer Graphics/04. Geometry Shader/mesh.h
@@ -229,7 +229,7 @@ private:
 		const ComPtr<ID3D12GraphicsCommandList>& commandList, const wstring& fileName) override;
 
 private:
-	vector<vector<BYTE>> m_height;
+	vector<vector<FLOAT>> m_height;
 	INT m_length;
 	FLOAT m_grid;
 };
